requirements.txt
streamlit
pillow
opencv-python
numpy
pytesseract
PyYAML
